backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных и ORM
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Валидация данных
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis и кеширование
redis==5.0.1
aioredis==2.0.1

# Асинхронные задачи
celery==5.3.4

# HTTP клиент
httpx==0.25.2
aiohttp==3.9.1

# Утилиты
python-dotenv==1.0.0
email-validator==2.1.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Разработка
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Безопасность
bandit==1.7.5
safety>=2.4.0

# Tinkoff Invest API SDK (production)
tinkoff-investments
